fixed bug caused by size_t; sorted edges in wrapper; refactored wrapper in its own struct; changed name of package functions file

+++ src/model.cpp
@@ -24,7 +24,7 @@ double model::penalized_information(variable x, variable y)
   return info;
 }
 
-int model::get_df() const
+size_t model::get_df() const
 {
   return df;
 }
